third_party/blink/web_tests/platform/generic/fast/multicol/image-inside-nested-blocks-with-border-expected.txt

Test that getBoundingClientRect works as expected on a float in the second column

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS result.left is 300
PASS result.top is 0
PASS successfullyParsed is true

TEST COMPLETE
